driver-drowsiness - Copy/requirements.txt
# Core Matrix
numpy==1.26.4
opencv-python==4.8.1.78
scipy==1.11.4
dlib==19.24.2
imutils==0.5.1

# AI & Voice Engine
google-generativeai==0.8.6
pvporcupine==3.0.0
PyAudio==0.2.14
pyttsx3==2.99
SpeechRecognition==3.14.4
requests==2.31.0

# Modern GUI
customtkinter==5.2.2
Pillow==12.0.0

# OS Utils
pygame==2.5.2


annotated-types==0.7.0
blis==1.3.3
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
comtypes==1.4.13
confection==0.1.5
customtkinter==5.2.2
cymem==2.0.13
darkdetect==0.8.0
dlib==19.24.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85   
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
imutils==0.5.1
Jinja2==3.1.6
langcodes==3.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numpy==1.26.4
opencv-python==4.8.1.78
packaging==25.0
pillow==12.0.0
preshed==3.0.12
proto-plus==1.27.0
protobuf==5.29.5
pvporcupine==3.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.5.2
Pygments==2.19.2
pyparsing==3.2.5
pypiwin32==223
pyttsx3==2.99
pywin32==311
requests==2.31.0
rich==14.2.0
rsa==4.9.1
scipy==1.11.4
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.4
srsly==2.5.2
thinc==8.3.10
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.1
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
